Refactor: SessionKey -> SessionAttributes
- 상수 클래스의 이름 변경: SessionKey -> SessionAttributes
- 위치 변경: authentication.constant -> authentication.session
