+++ src/main/java/com/uranus/taskmanager/api/global/audit/SessionAuditorAware.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.stereotype.Component;
 
-import com.uranus.taskmanager.api.security.authentication.constant.SessionKey;
+import com.uranus.taskmanager.api.security.authentication.session.SessionAttributes;
 
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -17,7 +17,7 @@ public class SessionAuditorAware implements AuditorAware<String> {
 
 	@Override
 	public Optional<String> getCurrentAuditor() {
-		String loginId = (String)session.getAttribute(SessionKey.LOGIN_MEMBER_LOGIN_ID);
+		String loginId = (String)session.getAttribute(SessionAttributes.LOGIN_MEMBER_LOGIN_ID);
 		return Optional.ofNullable(loginId);
 	}
 }
